+++ changelog.md
@@ -20,6 +20,7 @@ For example:
 * Added multi-cell clearing via backspace key (delete on Mac).
 * Added API exception registry that allows plugins to provide custom exception mappings for the REST API.
 * Added formula round and int functions. [#891](https://gitlab.com/bramw/baserow/-/issues/891)
+* Link to table field can now link rows in the same table. [#798](https://gitlab.com/bramw/baserow/-/issues/798)
 
 ### Bug Fixes
 
@@ -41,6 +42,7 @@ For example:
 * Fix lookup field conversions deleting all of its old field dependencies. [#1036](https://gitlab.com/bramw/baserow/-/issues/1036)
 
 ### Breaking Changes
+
 
 ## Released (2022-06-09 1.10.1)